bpf: tests: condense TC Dynamic-NAT LB test
Run the IPv4 and IPv6 test from a single file.
